added Tensor<float> to Tensor<double> broadcasting

+++ tests/tensors/test_broadcast.cpp
@@ -1,7 +1,6 @@
 #include "../../statkitcpp/core/tensor.h"
 
 #include <catch2/catch_test_macros.hpp>
-#include <iostream>
 
 namespace statkitcpp {
 
@@ -18,7 +17,7 @@ TEST_CASE("Simple broadcast") {
     }
     {
         Tensor<float> a({2, 1});
-        Tensor<float> b({1});
+        Tensor<double> b({1});
         REQUIRE(a.BroadcastableTo(b));
     }
 }
